requirements.txt
faiss-cpu
openai==0.28
python-dotenv
scikit-learn
sentence-transformers


